+++ examples/example-paper-i.c
@@ -16,14 +16,14 @@
 
 #if 1
 static struct instr_paper program[] = {
-	{'b', BLANK,   "P0,R", 'c'},
-	{'c', BLANK,   "R",    'e'},
-	{'e', BLANK,   "P1,R", 'f'},
-	{'f', BLANK,   "R",    'b'},
+	{'b', NONE,   "P0,R", 'c'},
+	{'c', NONE,   "R",    'e'},
+	{'e', NONE,   "P1,R", 'f'},
+	{'f', NONE,   "R",    'b'},
 };
 #else
 static struct instr_paper program[] = {
-	{'b', BLANK, "P0",     'b'},
+	{'b', NONE,  "P0",     'b'},
 	{'b', '0',   "R,R,P1", 'b'},
 	{'b', '1',   "R,R,P0", 'b'},
 };