+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Release history
 * __v1.0.1 (2020-xx-xx)__
 
  Changes:
+   - Use NONE instead of BLANK for paper programs
 
  Bug Fixes: